Week 1/Day 2/pdf_rag/requirements.txt
PyPDF2
numpy
openai
python-dotenv